voice-agent-pipecat/requirements.txt
pipecat-ai[deepgram,google,cartesia,silero,webrtc]==0.0.84
python-dotenv~=1.0.1
httpx>=0.27.0
aiohttp>=3.8.0
